function/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.23.1
google-cloud-storage==2.18.2